修改错误

+++ src/448C++20explicit关键字.cpp
@@ -15,7 +15,7 @@ struct S {
 int main() {
     //explicit 说明符可以与常量表达式一同使用。函数在且只会在该常量表达式求值为 true 时是显式的
     S<int>s;
-    std::cout << s << '\n';//int,因为operator int的explicit常量表达式为true，所以无法隐式转换，重载走double
+    std::cout << s << '\n';//double,因为operator int的explicit常量表达式为true，所以无法隐式转换，重载走double
     S<double>s2;
-    std::cout << s2 << '\n';//double
+    std::cout << s2 << '\n';//int
 }
